Add default AudioAdapter when no adapter passed in
